[main] Create Status Code for each endpoint | format

+++ visacards-api/src/main/java/com/afonsoqueiros/springbootinduction/visacardsapi/domain/services/VisaCardService.java
@@ -4,6 +4,7 @@ import com.afonsoqueiros.springbootinduction.visacardsapi.domain.VisaCard;
 import com.afonsoqueiros.springbootinduction.visacardsapi.dtos.CreateVisaCard;
 import com.afonsoqueiros.springbootinduction.visacardsapi.dtos.GetVisaCard;
 import com.afonsoqueiros.springbootinduction.visacardsapi.dtos.UpdateVisaCard;
+import com.afonsoqueiros.springbootinduction.visacardsapi.exceptions.GlobalExceptions;
 import com.afonsoqueiros.springbootinduction.visacardsapi.repository.VisaCardRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,26 +23,42 @@ public class VisaCardService {
         return VisaCardMapper.INSTANCE.mapVisaCardToGetVisaCard(visaCard);
     }
 
-    public boolean createVisaCard(CreateVisaCard createVisaCard){
+    public void createVisaCard(CreateVisaCard createVisaCard){
 
         VisaCard visaCard = VisaCardMapper.INSTANCE.mapCreateVisaCardToVisaCard(Optional.ofNullable(createVisaCard));
-        this.visaCardRepository.save(visaCard);
-
-        return true;
+        if(visaCard != null){
+            if(this.visaCardRepository.findVisaCardByCardNumber(visaCard.getCardNumber())!= null){
+                throw new GlobalExceptions.ResourceAlreadyOnDBException();
+            }
+            else{
+                this.visaCardRepository.save(visaCard);
+            }
+        }
+        else {
+            throw new GlobalExceptions.PayloadInvalidException();
+        }
     }
 
-    public boolean updateVisaCard(UpdateVisaCard updateVisaCard, Long id){
+    public void updateVisaCard(UpdateVisaCard updateVisaCard, Long id){
 
         Optional<VisaCard> visaCardToUpdate = this.visaCardRepository.findById(id);
-        VisaCard visaCard = VisaCardMapper.INSTANCE.mapUpdateVisaCardToVisaCard(Optional.ofNullable(updateVisaCard),visaCardToUpdate);
-
-        this.visaCardRepository.save(visaCard);
-
-        return true;
+        if(visaCardToUpdate.isPresent()){
+            VisaCard visaCard = VisaCardMapper.INSTANCE.mapUpdateVisaCardToVisaCard(Optional.ofNullable(updateVisaCard),visaCardToUpdate);
+            this.visaCardRepository.save(visaCard);
+        }
+        else{
+            throw new GlobalExceptions.ResourceNotFoundException();
+        }
     }
 
-    public boolean deleteVisaCard(Long id){
-        this.visaCardRepository.deleteById(id);
-        return true;
+    public void deleteVisaCard(Long id){
+        Optional<VisaCard> visaCardToDelete = this.visaCardRepository.findById(id);
+        if(visaCardToDelete.isPresent()){
+            this.visaCardRepository.deleteById(id);
+        }
+        else{
+            throw new GlobalExceptions.ResourceNotFoundException();
+        }
     }
+
 }
